+++ Example macOS/ViewController.swift
@@ -41,8 +41,8 @@ class ViewController: NSViewController {
         let parser = FeedParser(URL: feedURL)
 
         // Parse asynchronously, not to block the UI.
-        parser.parseAsync { [weak self] (result) in
-            guard let self = self else { return }
+        Task {
+            let result = parser.parseAsync()
             switch result {
             case .success(let feed):
                 // Grab the parsed feed directly as an optional rss, atom or json feed object
@@ -57,7 +57,7 @@ class ViewController: NSViewController {
                 // }
                 
                 // Then back to the Main thread to update the UI.
-                DispatchQueue.main.async {
+                MainActor.run {
                     self.feedItemsTableView.reloadData()
                 }