+++ Tests/ContentTests.swift
@@ -60,11 +60,10 @@ class ContentTests: BaseTestCase {
             let parser = FeedParser(URL: URL)
             
             // When
-            parser.parseAsync { (result) in
-                
+            Task {
+                _ = await parser.parseAsync()
                 // Then
                 expectation.fulfill()
-                
             }
             
             self.waitForExpectations(timeout: self.timeout, handler: nil)