+++ Tests/DublinCoreTests.swift
@@ -92,11 +92,10 @@ class DublinCoreTests: BaseTestCase {
             let parser = FeedParser(URL: URL)
             
             // When
-            parser.parseAsync{ (result) in
-                
+            Task {
+                _ = await parser.parseAsync()
                 // Then
                 expectation.fulfill()
-                
             }
             
             self.waitForExpectations(timeout: self.timeout, handler: nil)
@@ -170,11 +169,10 @@ class DublinCoreTests: BaseTestCase {
             let parser = FeedParser(URL: URL)
             
             // When
-            parser.parseAsync{ (result) in
-                
+            Task {
+                _ = await parser.parseAsync()
                 // Then
                 expectation.fulfill()
-                
             }
             
             self.waitForExpectations(timeout: self.timeout, handler: nil)