+++ Tests/MediaTests.swift
@@ -262,11 +262,10 @@ class MediaTests: BaseTestCase {
             let parser = FeedParser(URL: URL)
             
             // When
-            parser.parseAsync { (result) in
-                
+            Task {
+                _ = await parser.parseAsync()
                 // Then
                 expectation.fulfill()
-                
             }
             
             self.waitForExpectations(timeout: self.timeout, handler: nil)
@@ -479,11 +478,10 @@ class MediaTests: BaseTestCase {
             let parser = FeedParser(URL: URL)
             
             // When
-            parser.parseAsync { (result) in
-                
+            Task {
+                _ = await parser.parseAsync()
                 // Then
                 expectation.fulfill()
-                
             }
             
             self.waitForExpectations(timeout: self.timeout, handler: nil)