+++ Tests/RDFTests.swift
@@ -73,11 +73,10 @@ class RDFTests: BaseTestCase {
             let parser = FeedParser(URL: URL)
             
             // When
-            parser.parseAsync { (result) in
-                
+            Task {
+                _ = await parser.parseAsync()
                 // Then
                 expectation.fulfill()
-                
             }
             
             self.waitForExpectations(timeout: self.timeout, handler: nil)